# docs/openapi.v1.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "MyBabyNotes API v1",
        "version": "1.0",
        "description": "The MyBabyNotes public API. Authenticate with a personal access token from Settings \u2192 API access (`Authorization: Bearer <token>`). See docs/integrations.md for scopes, conventions, and examples."
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/entries/{id}": {
            "get": {
                "operationId": "entry.show",
                "tags": [
                    "Entry"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`EntryResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EntryResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "entry.update",
                "tags": [
                    "Entry"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateEntryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`EntryResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EntryResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "entry.destroy",
                "summary": "Tombstone, never a hard delete \u2014 the row keeps syncing as deleted",
                "tags": [
                    "Entry"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`EntryResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EntryResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/timer": {
            "get": {
                "operationId": "timer.show",
                "tags": [
                    "Timer"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "timer": {
                                            "type": [
                                                "object",
                                                "null"
                                            ],
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                },
                                                "type": {
                                                    "type": "string"
                                                },
                                                "started_at": {
                                                    "type": "integer"
                                                },
                                                "user_id": {
                                                    "type": "integer"
                                                },
                                                "baby_id": {
                                                    "type": [
                                                        "integer",
                                                        "null"
                                                    ]
                                                }
                                            },
                                            "required": [
                                                "id",
                                                "type",
                                                "started_at",
                                                "user_id",
                                                "baby_id"
                                            ]
                                        },
                                        "timers": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "server_time": {
                                            "type": "integer"
                                        }
                                    },
                                    "required": [
                                        "timer",
                                        "timers",
                                        "server_time"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "put": {
                "operationId": "timer.store",
                "tags": [
                    "Timer"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StartTimerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "timer": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                },
                                                "type": {
                                                    "type": "string"
                                                },
                                                "started_at": {
                                                    "type": "integer"
                                                },
                                                "user_id": {
                                                    "type": "integer"
                                                },
                                                "baby_id": {
                                                    "type": [
                                                        "integer",
                                                        "null"
                                                    ]
                                                }
                                            },
                                            "required": [
                                                "id",
                                                "type",
                                                "started_at",
                                                "user_id",
                                                "baby_id"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "timer"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "timer.destroy",
                "tags": [
                    "Timer"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "which timer to stop; omitted = your newest (the legacy singular slot)",
                        "schema": {
                            "type": "string",
                            "maxLength": 64
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "stopped": {
                                            "type": [
                                                "object",
                                                "null"
                                            ],
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                },
                                                "type": {
                                                    "type": "string"
                                                },
                                                "started_at": {
                                                    "type": "integer"
                                                },
                                                "user_id": {
                                                    "type": "integer"
                                                },
                                                "baby_id": {
                                                    "type": [
                                                        "integer",
                                                        "null"
                                                    ]
                                                }
                                            },
                                            "required": [
                                                "id",
                                                "type",
                                                "started_at",
                                                "user_id",
                                                "baby_id"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "stopped"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "description": "Personal access token from Settings \u2192 API access. Scopes: `profile:read` (Read your profile and household members), `children:read` (Read the children in your log), `entries:read` (Read log entries), `entries:write` (Create, edit, and delete log entries), `timer:read` (See the running timer), `timer:write` (Start and stop timers), `mcp` (Connect AI assistants via MCP).",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "EntryResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "baby_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "type": {
                        "type": "string"
                    },
                    "t": {
                        "type": "integer"
                    },
                    "detail": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "deleted": {
                        "type": "boolean"
                    },
                    "rev": {
                        "type": "integer"
                    }
                },
                "required": [
                    "id",
                    "baby_id",
                    "user_id",
                    "type",
                    "t",
                    "detail",
                    "deleted",
                    "rev"
                ],
                "title": "EntryResource"
            },
            "StartTimerRequest": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "nurse",
                            "pump",
                            "sleep",
                            "tummy"
                        ]
                    },
                    "baby_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "started_at": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "description": "Backdate the start (epoch ms), for a timer started after the baby already went down. Clamped to now at the latest and a day back at most.",
                        "minimum": 0
                    }
                },
                "required": [
                    "type"
                ],
                "title": "StartTimerRequest"
            },
...
            "UpdateEntryRequest": {
                "type": "object",
                "description": "Patch one entry: unspecified fields keep their stored values.",
                "properties": {
                    "type": {
                        "type": "string",
                        "maxLength": 20
                    },
                    "t": {
                        "type": "integer"
                    },
                    "detail": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 100
                    },
                    "baby_id": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "description": "in-household only; sending it re-homes the entry, omitting it never does"
                    },
                    "deleted": {
                        "type": "boolean"
                    }
                },
                "title": "UpdateEntryRequest"
            }
        },
        "responses": {
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}